backend/requirements.txt
fastapi
uvicorn
yt-dlp
requests
pydantic
python-dotenv
assemblyai
openai
